refactor(tpcc): extract shared core into tpcc_common.ts
procs.ts and tx.ts duplicated ~600 lines: config constants, compliance
metrics, the nine seeded load specs, the load + population-validation
logic, and the scenario/options. Move all of it into tpcc_common.ts and
have both variants import it; each keeps only its transaction bodies,
dialect branches (renderDDL, IS_*), retry strategy, and handleSummary.
tx.ts keeps its three extra client-observable counters (remote-line, BC).

Net ~628 fewer lines (2863 -> 2235). Behaviour unchanged: same seeds =>
byte-identical load. Verified on real pg/mysql/ydb (large-vm): procs
two-run + tx single, validate_population green on every dialect.
